requirements.txt
asgiref==3.6.0
Django==4.1.5
djangorestframework==3.14.0
factory-boy==3.2.1
Faker==16.6.1
python-dateutil==2.8.2
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
numpy==1.24.1
pandas==1.5.3
